+++ src/Thunderegg/Poisson/StencilHelper2d.h
@@ -273,9 +273,9 @@ class CoarseSH : public StencilHelper
 	int size(int i)
 	{
 		if (i == 0 || i == n - 1) {
-			return end_coeffs.size();
+			return (int) end_coeffs.size();
 		} else {
-			return mid_coeffs.size();
+			return (int) mid_coeffs.size();
 		}
 	}
 	double *coeffs(int i)
@@ -342,7 +342,7 @@ class FineSH : public StencilHelper
 		double h       = 0;
 		int    idx     = d.global_index * d.ns[0] * d.ns[1];
 		int    nbr_idx = d.getCoarseNbrInfo(s).global_index * d.ns[0] * d.ns[1];
-		end            = d.getCoarseNbrInfo(s).orth_on_coarse.getIndex() == 1;
+		end            = (int) d.getCoarseNbrInfo(s).orth_on_coarse.getIndex() == 1;
 		if (s == Side<2>::west()) {
 			h         = d.spacings[1];
 			start     = idx;
